professional viewing pages and order validations
The items in your cart can now be ordered.
You can now locate all of the orders on a professional menu and verify that they are complete. Additionally, you have an analytics page that shows how much beer you have validated on a daily, weekly, monthly, quarterly, and annual basis.