Payment options in the cart, order status updates, and analytics-optimized filters
You must now enter one of the following payment methods in the cart: Bancontact or PayPal Cash. A payment simulation appears when you click on PayPal.

You may now view the client's selected payment method and whether an order has been paid for or not in the order in progress page.
In order to maintain track of the order, I added the order date so that the most recent orders would show up underneath the older ones.

You can now see the beer precisely filtered by day, week, month, semester, and year with the precise number of beer orders for that time period. This is because, for the analytics, when an order is confirmed, a corresponding date is established with the matching amount of beer ordered.
Lastly, you can view past years using the year filter to see how many sales you made over that period.