Small changes after exchange with client
To prevent accidental clicking, I first repositioned the cart's remove all button to the top right corner for a more understated appearance.

If you click on the beer of the moment or the type of the moment on the homepage, you will be dynamically redirected to the order screen with the appropriate filter.

If you want to cancel an order that has already been paid for on the order in process screen, the app will notify you that you will need to issue a refund. After that, the order will be marked as "removed and reimbursed."
Apart from that, removing an order will cause it to shift to a section called "removed_orders" rather than being erased from the database.

I then updated the BeerDetails panel with additional information.

In order to determine whether there were price adjustments and when your revenue was at its peak, I have included the real price of the beer and the total revenue for that beer per filter. Additionally, each filter has a total revenue of the beers for that filter.