Last fixes
Each page now has a back button to return to the previous one.

The Firebase authenticator is now in charge of managing the professional login.

In order to better serve the needs of the customers, the chatbot now has access to my form responses, the type of beers, and the beers themselves.